Wrap listings in a special section
This was the best way I've found to enclose listings in a way that provides
control over the contents, shows up in tables, and marks the start and end of
the referenced entity semantically. Enclosing the whole thing in a float didn't
work because the splittext environment doesn't seem to work quite right: if you
set text as splittable, it always takes a whole page.