Update math implementation
This is a bit better because it allows captions but it's probably
not what we want just yet. Captions are treated as numbers rather
than a title under the equation. Maybe equations need to be wrapped
in a custom enumeration environment to be able to label them, but
I will need to review the semantics of equations in Org to know